requirements.txt
asgiref==3.7.2
autopep8==1.5.7
dj-database-url==0.5.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.25.0
psycopg==3.1.9
#psycopg2==2.9.6
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
Unipath==1.1
whitenoise==5.3.0
